+++ changelog.md
@@ -7,12 +7,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.1.0]
+
 ### Added
 
  - The functions OpenEventLog, CloseEventLog and ReadEventLog were added to
    advapi.
  - The sample uptime_evt.py demonstrates a way how these functions can be used.
  - The module msi and the accompanying sample exbinmsi.py were added.
+ - The functions get_directory_info and enum_directory_info were added to
+   ntdll. The sample listpipes.py demonstrates a way how these functions can
+   be used.
 
 ## [1.0.0]